+++ pybamm/models/full_battery_models/base_battery_model.py
@@ -143,6 +143,9 @@ class BatteryModelOptions(pybamm.FuzzyDict):
                     .. math::
                         \\eta_r = \\frac{F}{RT}
                         * (\\phi_s - \\phi_e - U - R_{sei} * L_{sei} * \\frac{I}{aL})
+            * "SEI on cracks" : str
+                Whether to include SEI growth on particle cracks, can be "false"
+                (default) or "true".
             * "SEI porosity change" : str
                 Whether to include porosity change due to SEI formation, can be "false"
                 (default) or "true".
@@ -250,6 +253,7 @@ class BatteryModelOptions(pybamm.FuzzyDict):
                 "ec reaction limited",
             ],
             "SEI film resistance": ["none", "distributed", "average"],
+            "SEI on cracks": ["false", "true"],
             "SEI porosity change": ["false", "true"],
             "stress-induced diffusion": ["false", "true"],
             "surface form": ["false", "differential", "algebraic"],
@@ -287,14 +291,21 @@ class BatteryModelOptions(pybamm.FuzzyDict):
         # The "SEI film resistance" option will still be overridden by extra_options if
         # provided
 
-        # Change the default for particle mechanics based on which LAM option is
-        # provided
-        # return "none" if option not given
-        lam_option = extra_options.get("loss of active material", "none")
-        if "stress-driven" in lam_option or "stress and reaction-driven" in lam_option:
-            default_options["particle mechanics"] = "swelling only"
+        # Change the default for particle mechanics based on which SEI on cracks option
+        # is provided
+        # return "false" if option not given
+        SEI_cracks_option = extra_options.get("SEI on cracks", "false")
+        if SEI_cracks_option == "true":
+            default_options["particle mechanics"] = "swelling and cracking"
         else:
-            default_options["particle mechanics"] = "none"
+            # Change the default for particle mechanics based on which LAM option is
+            # provided
+            # return "none" if option not given
+            LAM_opt = extra_options.get("loss of active material", "none")
+            if "stress-driven" in LAM_opt or "stress and reaction-driven" in LAM_opt:
+                default_options["particle mechanics"] = "swelling only"
+            else:
+                default_options["particle mechanics"] = "none"
         # The "particle mechanics" option will still be overridden by extra_options if
         # provided
 
@@ -448,6 +459,10 @@ class BatteryModelOptions(pybamm.FuzzyDict):
                 raise pybamm.OptionError(
                     f"X-lumped thermal submodels do not yet support {n}D "
                     "current collectors in a half-cell configuration"
+                )
+            elif options["SEI on cracks"] == "true":
+                raise NotImplementedError(
+                    "SEI on cracks not yet implemented for half-cell models"
                 )
 
         # Check options are valid