+++ pybamm/models/full_battery_models/lithium_ion/base_lithium_ion_model.py
@@ -164,11 +164,13 @@ class BaseModel(pybamm.BaseBatteryModel):
         # Different way of measuring LLI but should give same value
         LLI_sei = self.variables["Loss of lithium to SEI [mol]"]
         if self.half_cell:
+            LLI_sei_cracks = pybamm.Scalar(0)
             LLI_pl = pybamm.Scalar(0)
         else:
+            LLI_sei_cracks = self.variables["Loss of lithium to SEI on cracks [mol]"]
             LLI_pl = self.variables["Loss of lithium to lithium plating [mol]"]
 
-        LLI_reactions = LLI_sei + LLI_pl
+        LLI_reactions = LLI_sei + LLI_sei_cracks + LLI_pl
         self.variables.update(
             {
                 "Total lithium lost to side reactions [mol]": LLI_reactions,
@@ -212,6 +214,8 @@ class BaseModel(pybamm.BaseBatteryModel):
                 "Total lithium in negative electrode [mol]",
                 "Loss of lithium to lithium plating [mol]",
                 "Loss of capacity to lithium plating [A.h]",
+                "Loss of lithium to SEI on cracks [mol]",
+                "Loss of capacity to SEI on cracks [A.h]",
             ]
 
         self.summary_variables = summary_variables
@@ -238,8 +242,22 @@ class BaseModel(pybamm.BaseBatteryModel):
             self.submodels["sei"] = pybamm.sei.ConstantSEI(self.param, self.options)
         else:
             self.submodels["sei"] = pybamm.sei.SEIGrowth(
-                self.param, reaction_loc, self.options
+                self.param, reaction_loc, self.options, cracks=False
             )
+        # Do not set "sei on cracks" submodel for half-cells
+        # For full cells, "sei on cracks" submodel must be set, even if it is zero
+        if reaction_loc != "interface":
+            if (
+                self.options["SEI"] in ["none", "constant"]
+                or self.options["SEI on cracks"] == "false"
+            ):
+                self.submodels["sei on cracks"] = pybamm.sei.NoSEI(
+                    self.param, self.options, cracks=True
+                )
+            else:
+                self.submodels["sei on cracks"] = pybamm.sei.SEIGrowth(
+                    self.param, reaction_loc, self.options, cracks=True
+                )
 
     def set_lithium_plating_submodel(self):
         if self.options["lithium plating"] == "none":
@@ -261,7 +279,9 @@ class BaseModel(pybamm.BaseBatteryModel):
         for domain in ["Negative", "Positive"]:
             crack = getattr(self.options, domain.lower())["particle mechanics"]
             if crack == "none":
-                pass
+                self.submodels[
+                    domain.lower() + " particle mechanics"
+                ] = pybamm.particle_mechanics.NoMechanics(self.param, domain)
             elif crack == "swelling only":
                 self.submodels[
                     domain.lower() + " particle mechanics"